# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/Area/AcreToFoot2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/AcreToMile2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/AcreToYard2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Foot2ToAcre/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Foot2ToInch2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Inch2ToFoot2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Inch2ToYard2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Mile2ToAcre/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Rod2ToYard2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Area/Yard2ToAcre/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/Area/Yard2ToInch2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/Area/Yard2ToRod2/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
    "/Area/Circumference/{input}": {
      "get": {
        "tags": [
          "Area"
        ],
        "description": "",
        "parameters": [
          {
            "name": "input",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {}
      }
    },
...
  },
...
}